move deep_gemm_wrapper
