+++ python/sglang/srt/layers/quantization/fp8_utils.py
@@ -2,7 +2,7 @@ from typing import Callable, List, Optional, Tuple
 
 import torch
 
-from sglang.srt.layers.quantization import deep_gemm_wrapper
+from sglang.srt.layers import deep_gemm_wrapper
 from sglang.srt.layers.quantization.fp8_kernel import sglang_per_token_group_quant_fp8
 from sglang.srt.layers.quantization.mxfp4_tensor import MXFP4QuantizeUtil
 from sglang.srt.utils import is_sm100_supported, offloader
